proper changes from restserver to adminserver in Console, remove CmdAPI and RestAPI in favor of CommandClient and AdminAPI
