+++ tools/src/main/twirl/io/prediction/tools/console/adminserver.scala.txt
@@ -1,4 +1,4 @@
-Usage: pio restserver [--ip <value>] [--port <value>]
+Usage: pio adminserver [--ip <value>] [--port <value>]
 
   --ip <value>
       IP to bind to. Default: localhost
